+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- BarWidgetUI & HistogramWidgetUI: Add a skeleton for loading state [#674](https://github.com/CartoDB/carto-react/pull/674)
 - TablePagination fixes & DS application [#673](https://github.com/CartoDB/carto-react/pull/673)
 - Remove ReactDOMServer dependency and simplify avatar image fallback [#672](https://github.com/CartoDB/carto-react/pull/672)
 - Remove @mui/styles after dumping makeStyles [#670](https://github.com/CartoDB/carto-react/pull/670)